Updated the pom to use activemq 5.4.2-fuse-04-05 as the runtime version of the broker.
Added the ignoring of the hub,spoke and stomp prefixed files since these
are used only by the tests.

Correct the setAmqPort call in the test.  They were passing a string
instead of an integer.

Move the log4j.properties file to the test resources directory from
the main resources directory since this is only required during the
testing of the project.